[CodeSmell] Cleanup - fixed code smell for  CalendarDayEventIcon
