+++ src/components/Calendar/Calendar.js
@@ -2,7 +2,6 @@ import * as React from 'react';
 import {useState, useMemo} from 'react';
 import {Typography} from "@mui/material";
 import Calendar from 'react-calendar';
-import CardContent from '@mui/material/CardContent';
 import '../Calendar/calendar.css'
 import {BackgroundCard, EventCard, EventTypeCard, StudyRoomChatCard} from '../CustomMUIComponents/CustomCards';
 import PersistentDrawerLeft from "../NavDrawer/navDrawer";
@@ -17,7 +16,13 @@ import AddAPhotoIcon from '@mui/icons-material/AddAPhoto';
 import ImageUpload from "./ImageUpload";
 import CreateEvent from "./Event/CreateEvent";
 import AddIcon from "@mui/icons-material/Add";
-import {CalendarDayEventIcon} from "./Custom/CalendarComponent";
+import {
+    EventDisplay,
+    EventTypeHeader,
+    EventHeader,
+    isSameDate,
+    DayTile
+} from "./Custom/CalendarComponent";
 
 export default function CalendarView() {
     const [date, setDate] = useState(new Date());
@@ -56,12 +61,6 @@ export default function CalendarView() {
 
     }, [])
 
-    const isSameDate = (date1, date2) => (
-        date1.getFullYear() === date2.getFullYear()
-        && date1.getMonth() === date2.getMonth()
-        && date1.getDate() === date2.getDate()
-    )
-
     function setDates(d) {
         setDate(d);
     }
@@ -90,11 +89,13 @@ export default function CalendarView() {
             ? "academicHighlight"
             : ""
     );
-    
-    const calendarMonth = (
+
+    const calendarMonth =(
         <React.Fragment>
             <Calendar
-                tileContent={({ date }) => <DayTile key={date} day={date} />}
+
+                tileContent={({date}) => <DayTile key={date} day={date} events={events} categories={categories} />}
+                // tileContent={<CalendarDate date={date} events={events} categories={categories}/>}
                 tileClassName={AcademicEventsTile}
                 onChange={setDates}
                 value={date}
@@ -120,59 +121,6 @@ export default function CalendarView() {
             }/>
         </React.Fragment>
     )
-
-    const EventHeader = ({content})=>{
-        return(
-        <React.Fragment>
-            <CardContent>
-                <Typography color="white" fontWeight={500} style={{
-                    fontFamily: 'Roboto', alignItems: 'center', display: 'flex',
-                }}>
-                    {content}
-                </Typography>
-            </CardContent>
-        </React.Fragment>
-   ) }
-
-    const EventTypeHeader = ({content})=>{
-        return(
-            <React.Fragment>
-                <CardContent>
-                    <Typography color="white" fontWeight={500} style={{
-                        fontFamily: 'Roboto', alignItems: 'center', display: 'flex',
-                    }}>
-                        {content}
-                    </Typography>
-                </CardContent>
-            </React.Fragment>
-        ) }
-
-    const EventDisplay = ({startTime, endTime, header, description, startDate, link}) => {
-        const currentDate = new Date(startDate);
-        return (
-            <div style={{paddingBottom: 0, paddingTop: 0, width: '100%', display: 'flow'}}>
-
-                <div style={{display: 'inline-block', paddingLeft: '10px'}}>
-                    <Typography sx={{fontSize: 14}} color="text.secondary" gutterBottom>
-                        {startTime + "-" + endTime}, {currentDate.getFullYear()} - {currentDate.getMonth() < 9 ? '0' + (currentDate.getMonth() + 1) : currentDate.getMonth() + 1} - {currentDate.getDate() < 10 ? '0' + currentDate.getDate() : currentDate.getDate()}
-                    </Typography>
-                    <Typography data-test={header} sx={{mb: 1.5}} color="#000000" fontWeight={500} style={{fontFamily: 'Roboto'}}>
-                        {header}
-                    </Typography>
-
-                    <Typography variant="body2" color="text.secondary">
-                        {description}
-                    </Typography>
-
-                    <Typography variant="body2" color="text.secondary">
-                        <a href={link}>{link}</a>
-                    </Typography>
-
-                </div>
-
-            </div>
-        )
-    }
 
     const eventsDisplay = (
         <>
@@ -249,25 +197,6 @@ export default function CalendarView() {
             </div>
         </>
     )
-
-    const DayTile = ({ day }) => {
-        const eventsThisDay = events.filter((e) => {
-            const event = new Date(e.startDate);
-            return isSameDate(event, day)
-        });
-
-        let tileContent;
-
-        if (eventsThisDay.length < 1) {
-            tileContent = (<CalendarDayEventIcon key={day} eventType={"none"} categories={categories} />);
-        } else {
-            tileContent = eventsThisDay.map((e) => (
-                <CalendarDayEventIcon key={`day-${e._id}`} eventType={e.type} categories={categories}/>
-            ))
-        }
-
-        return tileContent;
-    }
 
     const calendarPageCards = useMemo(() => (
         <React.Fragment>
